pyproject.toml
[tool.poetry]
name = "afsbo"
version = "0.1.0"
description = ""
authors = ["VasilevGrigoriy <vasiliev-greg@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<3.12"
python-dotenv = "^1.0.1"


[tool.poetry.group.pyspark.dependencies]
findspark = "^2.0.1"
pyspark = "3.0.3"


[tool.poetry.group.research.dependencies]
matplotlib = "^3.0"
seaborn = "^0.13.0"
pandas = "^1.5"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.26.0"


[tool.poetry.group.logs.dependencies]
tqdm = "^4.66.1"


[tool.poetry.group.s3.dependencies]
boto3 = "^1.34.34"
botocore = "^1.34.34"


[tool.poetry.group.lint.dependencies]
black = "^24.1.1"
isort = "^5.13.2"

[tool.poetry.group.airflow.dependencies]
apache-airflow = "^2.8.1"
apache-airflow-providers-google = "^10.14.0"


[tool.poetry.group.click.dependencies]
click = "^8.1.7"


[tool.poetry.group.mlflow.dependencies]
mlflow = "^2.10.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
